Galaxy Config file
Applied to files where it was needed.
Fixed alot of warnings.

+++ galaxy/src/galaxy/components/Primitive2D.cpp
@@ -94,12 +94,12 @@ namespace galaxy
 			return m_vertexs;
 		}
 
-		const int Primitive2D::get_width() const noexcept
+		const float Primitive2D::get_width() const noexcept
 		{
 			return m_width;
 		}
 
-		const int Primitive2D::get_height() const noexcept
+		const float Primitive2D::get_height() const noexcept
 		{
 			return m_height;
 		}
